requirements.txt
PyQt5>=5.15.0
python-can>=3.3.0